requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-filter==2.2.0
djangorestframework==3.11.0
Faker==4.1.0
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
